third_party/WebKit/LayoutTests/http/tests/security/cross-frame-access-static-operation-return-wrapper.html

<script>
console.log("Tests that static operations return DOM wrappers which are associated with the caller's context, not associated with |this| which can be faked.");
// See also http://crbug.com/494640

var iframe = document.documentElement.appendChild(document.createElement('iframe'));
iframe.src = 'data:text/plain,hello';
iframe.onload = function() {
    var uri = IDBKeyRange.only.call(frames[0], 0).constructor.constructor('return location.href')();
    console.log(uri);
    if (window.testRunner) {
        testRunner.notifyDone();
    }
}

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}
</script>
