third_party/WebKit/LayoutTests/http/tests/security/cross-frame-access-static-operation-return-wrapper-expected.txt

CONSOLE MESSAGE: line 2: Tests that static operations return DOM wrappers which are associated with the caller's context, not associated with |this| which can be faked.
CONSOLE MESSAGE: line 9: http://127.0.0.1:8000/security/cross-frame-access-static-operation-return-wrapper.html
